pyproject.toml
[project]
name = "raggy-duckdb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Michael Booth", email = "michael@databooth.com.au" }
]
requires-python = ">=3.12"
dependencies = [
    "diskcache>=5.6.3",
    "duckdb>=1.3.2",
    "loguru>=0.7.3",
    "numpy>=2.3.2",
    "python-dotenv>=1.1.1",
    "together>=1.5.21",
]

[project.scripts]
raggy-duckdb = "raggy_duckdb:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
